parse double

+++ BordroKrediSorgu/Models/KrediSorgu.cs
@@ -8,7 +8,7 @@ namespace BordroKrediSorgu.Models
     public class KrediSorgu
     {
         public String HesapNo { get; set; }
-        public int MaasAdet { get; set; }
+        public double MaasAdet { get; set; }
         public double KrediTutar { get; set; }
     }
 }
